Simplify CreateEffectWarhead
Simplified and streamlined code, moved everything
that might be usable by other warheads to Warhead.
