+++ OpenRA.Mods.Common/Warheads/LeaveSmudgeWarhead.cs
@@ -25,9 +25,6 @@ namespace OpenRA.Mods.Common.Warheads
 		[Desc("Type of smudge to apply to terrain.")]
 		public readonly HashSet<string> SmudgeType = new HashSet<string>();
 
-		[Desc("How close to ground must the impact happen to spawn smudges.")]
-		public readonly WDist AirThreshold = new WDist(128);
-
 		[Desc("Percentual chance the smudge is created.")]
 		public readonly int Chance = 100;
