+++ OpenRA.Mods.Common/Warheads/Warhead.cs
@@ -9,25 +9,21 @@
  */
 #endregion
 
+using System;
 using System.Collections.Generic;
+using System.Linq;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Primitives;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.Common.Warheads
 {
+	[Flags]
 	public enum ImpactType
 	{
-		None,
-		Ground,
-		Air,
-		TargetHit
-	}
-
-	public enum ImpactTargetType
-	{
-		NoActor,
-		ValidActor,
-		InvalidActor
+		None = 0,
+		Terrain = 1,
+		Actor = 2,
 	}
 
 	[Desc("Base warhead class. This can be used to derive other warheads from.")]
@@ -41,6 +37,18 @@ namespace OpenRA.Mods.Common.Warheads
 
 		[Desc("What diplomatic stances are affected.")]
 		public readonly Stance ValidStances = Stance.Ally | Stance.Neutral | Stance.Enemy;
+
+		[Desc("Warheads detonating above this altitude that don't hit an actor directly will check target validity against the 'TargetTypeAir' target types.")]
+		public readonly WDist AirThreshold = new WDist(128);
+
+		[Desc("Target types to use when the warhead detonated at an altitude greater than 'AirThreshold'.")]
+		static readonly BitSet<TargetableType> TargetTypeAir = new BitSet<TargetableType>("Air");
+
+		[Desc("On which impact types the warhead is allowed to trigger (if there's a valid target).",
+			"Current options are Terrain and Actor. If only Terrain is listed, warhead will never trigger on actors.",
+			"If only Actor is listed, the warhead will never trigger on empty terrain.",
+			"If both are listed, the warhead will trigger if either any touched actor or the underlying terrain has a valid target type.")]
+		public readonly ImpactType ValidImpactTypes = ImpactType.Terrain | ImpactType.Actor;
 
 		[Desc("Can this warhead affect the actor that fired it.")]
 		public readonly bool AffectsParent = false;
@@ -94,6 +102,40 @@ namespace OpenRA.Mods.Common.Warheads
 				return false;
 
 			return true;
+		}
+
+		/// <summary>Checks if the warhead is valid against any actor or the terrain at impact position.</summary>
+		public virtual bool IsValidImpact(World world, WPos pos, Actor firedBy)
+		{
+			if (ValidImpactTypes == ImpactType.None)
+				return false;
+
+			var targetTile = world.Map.CellContaining(pos);
+			if (!world.Map.Contains(targetTile))
+				return false;
+
+			// Check whether the impact position overlaps with an actor's hitshape
+			var potentialVictims = world.FindActorsOnCircle(pos, WDist.Zero);
+			foreach (var victim in potentialVictims)
+			{
+				var activeShapes = victim.TraitsImplementing<HitShape>().Where(Exts.IsTraitEnabled);
+				if (!activeShapes.Any(i => i.Info.Type.DistanceFromEdge(pos, victim).Length <= 0))
+					continue;
+
+				// If we got here, the impact touches an actors' hitshape.
+				// If Actor is not a valid ImpactType, we return 'false' immediately regardless of terrain.
+				if (!ValidImpactTypes.HasFlag(ImpactType.Actor))
+					return false;
+				else if (IsValidAgainst(victim, firedBy))
+					return true;
+			}
+
+			if (!ValidImpactTypes.HasFlag(ImpactType.Terrain))
+				return false;
+
+			var dat = world.Map.DistanceAboveTerrain(pos);
+			var tileInfo = world.Map.GetTerrainInfo(targetTile);
+			return IsValidTarget(dat > AirThreshold ? TargetTypeAir : tileInfo.TargetTypes);
 		}
 	}
 }
